+++ internal/output/embed.go
